解决结束请求崩溃的Bug，以及请求结束崩溃的Bug 存在问题： 1. 偶尔因为数据AsyncTcpConnection 的数据列表崩溃

+++ IKSKinectDataCapture/KinectDataCaptureQueue.h
@@ -28,7 +28,7 @@ public:
 	virtual bool GetSkeleFrame(SkeletonFramePtr& frame);
 
 private:
-	const int MAX_BUFF_SIZE = 20;
+	const size_t MAX_BUFF_SIZE = 20;
 	std::mutex m_ColorMutex;
 	std::mutex m_DepthMutex;
 	std::mutex m_SkeletonMutex;
